% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Draw state map of the accidents in a year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numeric variable specifies the state of
interest by number.}

\item{year}{a numeric variable specifies the year of interest.}
}
\value{
This function prints state map of the accidents in a year.
        If the state doesn't exist then prints an error message.
        If there were no accidents in the state and year specified,
        then prints "no accidents to plot".
}
\description{
This is a function that draws the locations of the accidents in
the specified state in the specified year.
}
\examples{
fars_map_state(1, 2018)

}
