% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read dataset}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the name of the
csv dataset.}
}
\value{
This function prints the data frame form of a csv dataset.
        If the filename doesn't exist then prints an error message.
}
\description{
This is a function that loads the csv dataset if it exists, and
prints the data frame form of the dataset. You can specify the
name of the dataset (using the \code{filename} argument).
}
