% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Select months and years of interest}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of numbers giving the years of interest.}
}
\value{
This function prints months and years of interst. If datasets
        of one or more years do not exist in the datasset then generates
        a warning message.
}
\description{
This is a function that selects the months and years in specified years
from datasets of these years (using the \code{years} argument).
}
\examples{
fars_read_years(list(2017,2018))

}
