% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize datasets of interest}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of numbers giving the years of interest.}
}
\value{
This function prints the number of months in each years
        of interest.
}
\description{
This is a function that counts the number of months in each years
of interest (using the \code{years} argument).
}
