% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numeric variable giving the year of the file name.}
}
\value{
This function prints different dataset name by
        different year.
}
\description{
This is a function that makes the dataset names. You can insert
the year of the dataset into the file name
(using the \code{year} argument).
}
